+++ LiteOrm.Tests/Service/ServiceExceptionHookFlowTests.cs
@@ -1,3 +1,4 @@
+using Autofac.Extras.DynamicProxy;
 using LiteOrm.Common;
 using LiteOrm.DependencyInjection;
 using LiteOrm.Service;